Fix warning: CA1901 P/Invoke declarations should be portable    As it is declared in your code, the return type of P/Invoke 'RichTextBoxXhtmlSupportExtension.NativeMethods.SendMessage(HandleRef, int, IntPtr, ref RichTextBoxXhtmlSupportExtension.PARAFORMAT)' will be 4 bytes wide on 64-bit platforms. This is not correct, as the actual native declaration of this API indicates it should be 8 bytes wide on 64-bit platforms. Consult the MSDN Platform SDK documentation for help determining what data type should be used instead of 'int'. GitUI   RichTextBoxXhtmlSupportExtension.cs 345

+++ GitUI/Editor/RichTextBoxXhtmlSupportExtension.cs
@@ -342,7 +342,7 @@ namespace GitUI.Editor.RichTextBoxExtension
                 ref Point lParam);
 
             [DllImport("user32", CharSet = CharSet.Auto)]
-            internal static extern int SendMessage(HandleRef hWnd,
+            internal static extern IntPtr SendMessage(HandleRef hWnd,
                 int msg,
                 IntPtr wParam,
                 ref PARAFORMAT lp);
